Move run_rank0
