+++ BodoSQL/bodosql/tests/test_types/rest_test_harness.py
@@ -1,8 +1,8 @@
 from ddltest_harness import DDLTestHarness
 
 import bodosql
+from bodo.spawn.utils import run_rank0
 from bodo.tests.iceberg_database_helpers.utils import SparkAwsIcebergCatalog, get_spark
-from bodo.utils.utils import run_rank0
 
 
 class RestTestHarness(DDLTestHarness):
