+++ BodoSQL/bodosql/tests/test_types/utils.py
@@ -7,9 +7,9 @@ from pyiceberg.catalog import Catalog
 import bodo
 from bodo.io.iceberg.catalog import conn_str_to_catalog
 from bodo.mpi4py import MPI
+from bodo.spawn.utils import run_rank0
 from bodo.tests.utils import _test_equal_guard, gen_unique_table_id
 from bodo.tests.utils_jit import reduce_sum
-from bodo.utils.utils import run_rank0
 
 
 def gen_unique_id(name_prefix: str) -> str:
