+++ bodo/spawn/utils.py
@@ -4,11 +4,14 @@ This file should import JIT lazily to avoid slowing down non-JIT code paths.
 
 from __future__ import annotations
 
+import functools
 import logging
 import os
 import sys
+import traceback
 import typing as pt
 import uuid
+from collections.abc import Callable
 from enum import Enum
 from time import sleep
 
@@ -321,3 +324,45 @@ def _get_data_sample(data):
     raise ValueError(
         "_get_data_sample only supports DataFrame, Series and ArrowExtensionArray input"
     )
+
+
+def run_rank0(func: Callable, bcast_result: bool = True, result_default=None):
+    """
+    Utility function decorator to run a function on just rank 0
+    but re-raise any Exceptions safely on all ranks.
+    NOTE: 'func' must be a simple python function that doesn't require
+    any synchronization.
+    e.g. Using a bodo.jit function might be unsafe in this situation.
+    Similarly, a function that uses any MPI collective
+    operation would be unsafe and could result in a hang.
+
+    Args:
+        func: Function to run.
+        bcast_result (bool, optional): Whether the function should be
+            broadcasted to all ranks. Defaults to True.
+        result_default (optional): Default for result. This is only
+            useful in the bcase_result=False case. Defaults to None.
+    """
+
+    @functools.wraps(func)
+    def inner(*args, **kwargs):
+        comm = MPI.COMM_WORLD
+        result = result_default
+        err = None
+        # Run on rank 0 and catch any exceptions.
+        if comm.Get_rank() == 0:
+            try:
+                result = func(*args, **kwargs)
+            except Exception as e:
+                print("".join(traceback.format_exception(None, e, e.__traceback__)))
+                err = e
+        # Synchronize and re-raise any exception on all ranks.
+        err = comm.bcast(err)
+        if isinstance(err, Exception):
+            raise err
+        # Broadcast the result to all ranks.
+        if bcast_result:
+            result = comm.bcast(result)
+        return result  # type: ignore
+
+    return inner
