+++ bodo/tests/test_s3.py
@@ -397,7 +397,7 @@ def test_s3_parquet_write_seq(
     writing to s3_bucket_us_west_2 will check if the s3 auto region
     detection functionality works
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     request.getfixturevalue(bucket_fixture)
 
@@ -429,7 +429,7 @@ def test_s3_parquet_write_1D(minio_server_with_s3_envs, s3_bucket, test_df):
     """
     test s3 to_parquet in 1D distributed
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     def test_write(test_df):
         test_df.to_parquet("s3://bodo-test/test_df_bodo_1D.pq")
@@ -459,7 +459,7 @@ def test_s3_parquet_write_1D_var(minio_server_with_s3_envs, s3_bucket, test_df):
     """
     test s3 to_parquet in 1D var
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     def test_write(test_df):
         test_df.to_parquet("s3://bodo-test/test_df_bodo_1D_var.pq")
@@ -494,7 +494,7 @@ def test_s3_csv_write_seq(
     writing to s3_bucket_us_west_2 will check if the s3 auto region
     detection functionality works
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     request.getfixturevalue(bucket_fixture)
 
@@ -536,7 +536,7 @@ def test_s3_csv_write_1D(minio_server_with_s3_envs, s3_bucket, test_df):
     """
     test s3 to_csv in 1D distributed
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     def test_write(test_df):
         test_df.to_csv("s3://bodo-test/test_df_bodo_1D.csv", index=False, header=False)
@@ -580,7 +580,7 @@ def test_s3_csv_write_1D_var(minio_server_with_s3_envs, s3_bucket, test_df):
     """
     test s3 to_csv in 1D var
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     def test_write(test_df):
         test_df.to_csv(
@@ -624,7 +624,7 @@ def test_s3_csv_write_header_seq(minio_server_with_s3_envs, s3_bucket, test_df):
     """
     test s3 to_csv with header sequentially
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     def test_write(test_df):
         test_df.to_csv("s3://bodo-test/test_df_bodo_header_seq.csv", index=False)
@@ -651,7 +651,7 @@ def test_s3_csv_write_header_1D(minio_server_with_s3_envs, s3_bucket, test_df):
     """
     test s3 to_csv with header in 1D distributed
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     def test_write(test_df):
         test_df.to_csv("s3://bodo-test/test_df_bodo_header_1D.csv", index=False)
@@ -689,7 +689,7 @@ def test_s3_csv_write_header_1D_var(minio_server_with_s3_envs, s3_bucket, test_d
     """
     test s3 to_csv with header in 1D var
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     def test_write(test_df):
         test_df.to_csv("s3://bodo-test/test_df_bodo_header_1D_var.csv", index=False)
@@ -777,7 +777,7 @@ def test_s3_json_write_records_lines_seq(
     writing to s3_bucket_us_west_2 will check if the s3 auto region
     detection functionality works
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     request.getfixturevalue(bucket_fixture)
 
@@ -815,7 +815,7 @@ def test_s3_json_write_records_lines_1D(minio_server_with_s3_envs, s3_bucket, te
     """
     test s3 to_json(orient="records", lines=True) in 1D distributed
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     def test_write(test_df):
         test_df.to_json(
@@ -848,7 +848,7 @@ def test_s3_json_write_records_lines_1D_var(
     """
     test s3 to_json(orient="records", lines=True) in 1D var
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     def test_write(test_df):
         test_df.to_json(
@@ -880,7 +880,7 @@ def test_s3_parquet_read(minio_server_with_s3_envs, s3_bucket, test_df):
     read_parquet
     test the parquet file we just wrote sequentially
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     @run_rank0
     def write_table():
@@ -901,7 +901,7 @@ def test_s3_csv_read(minio_server_with_s3_envs, s3_bucket, test_df):
     read_csv
     test the csv file we just wrote sequentially
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     @run_rank0
     def write_table():
@@ -934,7 +934,7 @@ def test_s3_csv_read_header(minio_server_with_s3_envs, s3_bucket, test_df):
     read_csv with header and infer dtypes
     test the csv file we just wrote sequentially
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     @run_rank0
     def write_table():
@@ -1103,7 +1103,7 @@ def test_s3_json_read(minio_server_with_s3_envs, s3_bucket, test_df):
     read_json(orient="records", lines=True)
     test the json file we just wrote sequentially
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     @run_rank0
     def write_table():
