+++ bodo/tests/test_snowflake_write.py
@@ -1620,8 +1620,8 @@ def test_batched_write_agg(
         snowflake_writer_append_table,
         snowflake_writer_init,
     )
+    from bodo.spawn.utils import run_rank0
     from bodo.tests.utils_jit import reduce_sum
-    from bodo.utils.utils import run_rank0
 
     comm = MPI.COMM_WORLD
     col_meta = bodo.utils.typing.ColNamesMetaType(
@@ -1908,7 +1908,7 @@ def test_batched_write_nested_array(
     Test writing a table with a column of nested arrays to Snowflake
     and then reading it back
     """
-    from bodo.utils.utils import run_rank0
+    from bodo.spawn.utils import run_rank0
 
     if is_variant and write_type == "replace":
         pytest.skip("When replacing a table columns are never written as variant")
